Restructure state directories
