+++ frontend/src/app/app.component.ts
@@ -5,7 +5,7 @@ import { Observable, of, fromEvent, throwError } from 'rxjs';
 import { catchError, map, tap, first } from 'rxjs/operators';
 
 import { Store, select } from '@ngrx/store';
-import { updateStore } from './app.actions';
+// import { updateStore } from './app.actions';
 
 
 import { Socket } from 'ngx-socket-io';
