+++ frontend/src/models/models.ts
@@ -12,3 +12,8 @@ export interface Cell {
     row: number;
     column: number;
 }
+
+export interface UserData {
+    name: string;
+    id: string;
+}
